More restructuring of code. Main focus being push_forward_facet_normals
